Interface to register sensitive volumes and to cache its VMC VolID
We cannot rely on the TGeoVolume::GetNumber() function to get the volume
id. This seems to work with Geant3 but not for Geant4.
This commit is offering a new API that the developer/user can use to
register a sensitive volume and to query the volume ID from VMC behind
the scenes.

Application to MFT and V0 which where using GetNumber.